+++ Frameworks/MAWebView/Changes.md
@@ -1,5 +1,6 @@
 # Changes
 
 ## v1.0
+- Fixed the `isObserving` variable
 - Created an `isObserving` variable in the webView
 - Created a basic web view class